Учтен limitRate

+++ Downloader/Models/DownloadTask.cs
@@ -5,7 +5,7 @@ namespace Downloader.Models
 {
     public class DownloadTask
     {
-        public DownloadTask(int threadNumber, int limitRate, ConcurrentDictionary<string, IList<string>> links, string outputFolder)
+        public DownloadTask(int threadNumber, long limitRate, ConcurrentDictionary<string, IList<string>> links, string outputFolder)
         {
             Threads = threadNumber;
             LimitRate = limitRate;
@@ -14,7 +14,7 @@ namespace Downloader.Models
         }
 
         public int Threads { get; private set; }
-        public int LimitRate { get; private set; }
+        public long LimitRate { get; private set; }
         public ConcurrentDictionary<string, IList<string>> Links { get; private set; }
         public string OutputFolder { get; private set; }
     }
